Recomendacoes_whats/requirements.txt
Flask==2.2.2
requests==2.28.1
psycopg2==2.9.5
python-dotenv==1.0.0
